+++ app/src/main/java/woowacourse/shopping/data/datasource/product/local/LocalProductDataSource.kt
@@ -1,6 +1,7 @@
-package woowacourse.shopping.data.datasource.product
+package woowacourse.shopping.data.datasource.product.local
 
 import woowacourse.shopping.data.database.dao.product.ProductDao
+import woowacourse.shopping.data.datasource.product.ProductDataSource
 import woowacourse.shopping.data.model.DataProduct
 
 class LocalProductDataSource(private val dao: ProductDao) : ProductDataSource.Local {
